add support for Parallels Desktop VM images
Two new image types have been added:

1. parallels - this produces VM images with extension pvm.tgz that can be loaded directly into Parallels Desktop
2. vagrant_parallels - this produces a Vagrant box that works with parallels vagrant provider (http://parallels.github.io/vagrant-parallels/)

Just like vmdk and others we rely on qemu-img to convert raw images. Support for Parallels disk images was added to qemu-img in version 2.4.

I had to do a slight refactoring to add a concept of a VM_LAYOUT to support Parallels. Just like a disk layout, a vm layout defines how the VM is organized. Parallels Desktop requires a specific directory structure for its configuration and also for the disks. It not as simple as vmware vmx + vmdk side-by-side.

I also removed the box files from the actual image since there are not needed in /usr/share/oem.

Signed-off-by: Bassam Tabbara <bassam.tabbara@quantum.com>